Added interimAngle and interimDirection to event data
Currently the angle and the direction are always relative
to the first event. In practice that means, if one drags
100px to the right and then drags 50px back to the left,
the direction still says "right" even so one is actually
dragging to the left.

This commit adds two properties "interimAngle" and
"interimDirection" to the event object which give information
about the angle and direction relative to the last event instead
of relative to the first event.

Fixes #303